Fix logging

+++ wp-cli.sh
@@ -134,6 +134,7 @@ writeLog() {
 
 
 addDNSentry() {
+    echo addDNSentry | writeLog TRACE
     #body='{
     #    "login": "test-user",
     #    "nonce": "98475920834",
@@ -151,6 +152,7 @@ addDNSentry() {
 
 
 addVirtualHost() {
+    echo addVirtualHost | writeLog TRACE
     if [[ -d '/etc/apache2/sites-available' && ${WP_APACHE} -eq "1" ]];then
         cp ./server.virtualhost.conf "${WP_URL}.conf"
         replaceVarFile "${WP_URL}.conf" WP_URL "${WP_URL}"
@@ -164,6 +166,7 @@ addVirtualHost() {
 }
 
 replaceVarFile() {
+    echo replaceVarFile | writeLog TRACE
     file=$1
     find=$2
     replace=$3
@@ -172,6 +175,7 @@ replaceVarFile() {
 }
 
 getCert() {
+    echo getCert | writeLog TRACE
     domain=$1
     if [[ -d '/etc/letsencrypt' && ${WP_LETSENCRYPT} -eq 1 && ${WP_APACHE} -eq "0" ]]; then
         certbot certonly -d $domain  1> /tmp/stdout.log 2> /tmp/stderr.log; writeLog INFO < /tmp/stdout.log; writeLog ERR < /tmp/stderr.log
@@ -183,6 +187,7 @@ getCert() {
 }
 
 initWP() {
+    echo initWP | writeLog TRACE
     #docker ps -q -f name={container Name}
     if [[  $(docker-compose ps -a | grep ${WP_DB_NAME} -c ) -eq 0 ]]; then
         docker-compose up --no-start ${WP_DB_NAME} ${WP_NAME}  1> /tmp/stdout.log 2> /tmp/stderr.log; writeLog INFO < /tmp/stdout.log; writeLog ERR < /tmp/stderr.log
@@ -199,11 +204,13 @@ initWP() {
 }
 
 pluginsWP() {
+    echo pluginsWP | writeLog TRACE
     eval $wp_docker ${WP_CLI_NAME} wp plugin install --activate ${WP_PLUGINS[@]} 1> /tmp/stdout.log 2> /tmp/stderr.log; writeLog INFO < /tmp/stdout.log; writeLog ERR < /tmp/stderr.log
     eval $wp_docker ${WP_CLI_NAME} wp plugin auto-updates enable --all 1> /tmp/stdout.log 2> /tmp/stderr.log; writeLog INFO < /tmp/stdout.log; writeLog ERR < /tmp/stderr.log #awaiting update to implement this function
 }
 
 optionsWP() {
+    echo optionsWP | writeLog TRACE
     keylen=$(jq -r '.wp_options| length' env.json)
     #keyvalue
     for ((i=0;i<keylen;i++)); 
@@ -226,6 +233,7 @@ optionsWP() {
 }
 
 getWPoptions() {
+    echo getWPoptions | writeLog TRACE
     eval $wp_docker ${WP_CLI_NAME} wp option list --format=json --unserialize > wp_options.json 1> /tmp/stdout.log 2> /tmp/stderr.log; writeLog INFO < /tmp/stdout.log; writeLog ERR < /tmp/stderr.log
     eval $wp_docker ${WP_CLI_NAME} wp option list --format=yaml --unserialize > wp_options.yaml 1> /tmp/stdout.log 2> /tmp/stderr.log; writeLog INFO < /tmp/stdout.log; writeLog ERR < /tmp/stderr.log
     echo  "Exported to wp_options.json/yaml" | writeLog "INFO"
@@ -234,21 +242,27 @@ getWPoptions() {
 
 
 cleanWP() {
-        [[ ! -f  docker-compose.yml ]] && echo "WP is not initialized, please run --create first" && error=1 && exitScript
-        docker-compose rm --force --stop -v  1> /tmp/stdout.log 2> /tmp/stderr.log; writeLog INFO < /tmp/stdout.log; writeLog ERR < /tmp/stderr.log
-        docker volume prune --force  --filter label=com.docker.compose.project="$(basename $PWD)"  1> /tmp/stdout.log 2> /tmp/stderr.log; writeLog INFO < /tmp/stdout.log; writeLog ERR < /tmp/stderr.log
+    echo cleanWP | writeLog TRACE
+        if [[ ! -f  ./docker-compose.yml ]];then
+           echo "WP is not initialized, ignoring docker commands"
+        else
+           docker-compose rm --force --stop -v  1> /tmp/stdout.log 2> /tmp/stderr.log; writeLog INFO < /tmp/stdout.log; writeLog ERR < /tmp/stderr.log
+           docker volume prune --force  --filter label=com.docker.compose.project="$(basename $PWD)"  1> /tmp/stdout.log 2> /tmp/stderr.log; writeLog INFO < /tmp/stdout.log; writeLog ERR < /tmp/stderr.log
+        fi
         [[ ${WP_APACHE} -eq "1" && -f "./${WP_URL}.conf" ]] &&  rm ./"${WP_URL}.conf"
         [[ -f "./.WP_INITIALIZED" ]] && rm  "./.WP_INITIALIZED"
         [[ -f "./docker-compose.yml" ]] && rm "./docker-compose.yml"
         [[ ${WP_APACHE} -eq "1" && -f "/etc/apache2/sites-enabled/${WP_URL}.conf" ]] && rm  "/etc/apache2/sites-enabled/${WP_URL}.conf"
+        [[ ${WP_APACHE} -eq "1" && -f "/etc/apache2/sites-available/${WP_URL}.conf" ]] && rm  "/etc/apache2/sites-available/${WP_URL}.conf"
 }
 purge() {
+    echo purge | writeLog TRACE
     read -p "Are you sure (Y/n)" q
     if [[ $q -eq "Y" ]]; then
         cleanWP
         echo "You need to manually remove DNS entry"  | writeLog "INFO"
         read -p "Delete folder (Y/n)" f
-        if [[ $f -eq "Y" ]]; then
+        if [[ $f == "Y" ]]; then
             rm "../$(basename $PWD)"
         fi
     else 
@@ -261,6 +275,7 @@ purge() {
 
 
 update() {
+    echo update | writeLog TRACE
     echo "Updating wp-cli" | writeLog "INFO"
     docker pull wordpress:cli-php8.0 
     echo "Checking for wp updates"  | writeLog "INFO"
@@ -273,6 +288,7 @@ update() {
     eval $wp_docker ${WP_CLI_NAME} wp plugin update --all  1> /tmp/stdout.log 2> /tmp/stderr.log; writeLog INFO < /tmp/stdout.log; writeLog ERR < /tmp/stderr.log
 }
 createFiles() {
+    echo createFiles | writeLog TRACE
     addVirtualHost
     #backup old docker-compose.yml
     if [[ -f "./docker-compose.yml" ]]; then
@@ -292,25 +308,26 @@ exitScript() {
     for i in $(cat ./.env | cut -d ' ' -f2 |cut -d "=" -f 1);do
       unset $i
     done
+    [[ -f "./.env" ]] && rm  "./.env"
     IFS=IFS_OLD
     exit ${error}
 }
 
 main() {
+    echo main | writeLog TRACE
     if [[ -f './.WP_INITIALIZED' && $force -eq 0 ]]; then
         echo  "already initialised" | writeLog "INFO"
         exit 0
     elif [[ $force -eq 1 ]]; then
         cleanWP
         export force=0
-        main
-    elif [[ ! -f './.WP_INITIALIZED'  ]]; then
+        #main
+    fi
+    if [[ ! -f './.WP_INITIALIZED'  ]]; then
         #addDNSEntry
         createFiles
         initWP 
         touch ./.WP_INITIALIZED
-    else
-        echo "" 
     fi
 }
